Move the remote proxying into its own viewsets module
RemoteMixin and RemoteViewSet are the only thing the content viewsets package
still reaches back into content/api.py for, and neither belongs to the API
surface that file holds. They move to content/viewsets/remote.py, so the
package depends on nothing left behind.

The three NetworkClient patches follow the class to its new module.
