+++ kolibri/core/content/api.py
@@ -71,7 +71,6 @@ from kolibri.core.content.tasks import automatic_user_imported_resource_cleanup
 from kolibri.core.content.utils.cache import get_cache_key
 from kolibri.core.content.utils.cache import get_course_ids
 from kolibri.core.content.utils.cache import no_cache_on_method
-from kolibri.core.content.utils.cache import REMOTE_ETAG_CACHE_KEY
 from kolibri.core.content.utils.cache import remote_metadata_cache
 from kolibri.core.content.utils.cache import REMOTE_URL_PARAM
 from kolibri.core.content.utils.content_types_tools import (
@@ -93,13 +92,13 @@ from kolibri.core.content.utils.paths import get_local_content_storage_file_url
 from kolibri.core.content.utils.paths import get_v2_channel_lookup_url
 from kolibri.core.content.utils.search import get_available_metadata_labels
 from kolibri.core.content.utils.stopwords import stopwords_set
+from kolibri.core.content.viewsets.remote import RemoteMixin
 from kolibri.core.device.permissions import FromAppContextPermission
 from kolibri.core.discovery.utils.network.client import NetworkClient
 from kolibri.core.discovery.utils.network.errors import NetworkClientError
 from kolibri.core.discovery.utils.network.errors import NetworkLocationConnectionFailure
 from kolibri.core.discovery.utils.network.errors import NetworkLocationNotFound
 from kolibri.core.discovery.utils.network.errors import NetworkLocationResponseFailure
-from kolibri.core.discovery.utils.network.errors import ResourceGoneError
 from kolibri.core.discovery.well_known import CENTRAL_CONTENT_BASE_URL
 from kolibri.core.lessons.models import Lesson
 from kolibri.core.logger.models import ContentSessionLog
@@ -111,102 +110,9 @@ from kolibri.core.serializers import SplitTextField
 from kolibri.core.utils.pagination import ValuesViewsetCursorPagination
 from kolibri.core.utils.pagination import ValuesViewsetLimitOffsetPagination
 from kolibri.core.utils.pagination import ValuesViewsetPageNumberPagination
-from kolibri.utils.conf import OPTIONS
 from kolibri.utils.version import version_matches_range
 
 logger = logging.getLogger(__name__)
-
-
-class RemoteMixin:
-    def _should_proxy_request(self, request):
-        return REMOTE_URL_PARAM in request.GET
-
-    def _get_request_headers(self, request):
-        return {
-            "Accept": request.META.get("HTTP_ACCEPT"),
-            # Don't proxy client's accept encoding headers as it may include br for brotli
-            # that we cannot rely on having decompression for available on the server.
-            "Accept-Language": request.META.get("HTTP_ACCEPT_LANGUAGE"),
-            "Content-Type": request.META.get("CONTENT_TYPE"),
-            "If-None-Match": request.META.get("HTTP_IF_NONE_MATCH", ""),
-        }
-
-    def _get_response_headers(self, response):
-        headers = {}
-        header_names = ["Cache-Control", "Etag", "Expires", "Date", "Last-Modified"]
-        for header_name in header_names:
-            if header_name in response.headers:
-                headers[header_name] = response.headers[header_name.lower()]
-        return headers
-
-    def _cache_etag(self, baseurl, headers):
-        if "Etag" in headers:
-            cache_key = REMOTE_ETAG_CACHE_KEY.format(baseurl)
-            cache.set(cache_key, headers["Etag"], 3600)
-
-    def update_data(self, response_data, baseurl):
-        return response_data
-
-    def update_request_params(self, params, device_info):
-        # Hook for subclasses that proxy to a remote peer: rewrite query params
-        # so a request keeps working regardless of the Kolibri version exposing
-        # the endpoint being proxied to (e.g. remapping renamed params).
-        # Default is a no-op; device_info describes the peer being queried.
-        return params
-
-    def _hande_proxied_request(self, request):
-        full_path = request.get_full_path().split("?")[0]
-        remote_path = full_path.replace(
-            "/{}api/content/".format(
-                OPTIONS["Deployment"]["URL_PATH_PREFIX"].lstrip("/")
-            ),
-            "/api/public/v2/",
-        )
-        baseurl = request.GET[REMOTE_URL_PARAM]
-        qs = request.GET.copy()
-        del qs[REMOTE_URL_PARAM]
-        try:
-            client = NetworkClient.build_for_address(baseurl)
-        except NetworkLocationNotFound:
-            raise Http404("Remote resource not found")
-        qs = self.update_request_params(qs, client.device_info)
-        remote_url = remote_path
-        try:
-            response = client.get(
-                remote_url, params=qs, headers=self._get_request_headers(request)
-            )
-
-            # If Etag is set on the response we have returned here, any further Etag will not be modified
-            # by the django etag decorator, so this should allow us to transparently proxy the remote etag.
-            try:
-                content = self.update_data(response.json(), baseurl)
-            except ValueError:
-                content = response.content
-            headers = self._get_response_headers(response)
-            self._cache_etag(baseurl, headers)
-            return Response(
-                content,
-                status=response.status_code,
-                headers=headers,
-            )
-        except NetworkLocationResponseFailure as e:
-            if e.response.status_code == status.HTTP_404_NOT_FOUND:
-                raise Http404("Remote resource not found")
-            raise ResourceGoneError
-
-
-class RemoteViewSet(ReadOnlyValuesViewset, RemoteMixin):
-    def retrieve(self, request, pk=None):
-        if pk is None:
-            raise Http404
-        if self._should_proxy_request(request):
-            return self._hande_proxied_request(request)
-        return super().retrieve(request, pk=pk)
-
-    def list(self, request, *args, **kwargs):
-        if self._should_proxy_request(request):
-            return self._hande_proxied_request(request)
-        return super().list(request, *args, **kwargs)
 
 
 class ChannelThumbnailView(View):
